Nurobuckle/requirements.txt
streamlit>=1.24.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.13.0
httpx==0.27.0
twilio==8.12.0
python-dotenv==1.0.1
mcp-server==0.1.4
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit-folium==0.15.1
folium==0.15.1 